Adding support for conditional render plane settings.  These settings driven by a parameter name and a matching pattern.  SOHO will evaluate the parameter and compare it to the pattern.  Specific settings can be applied based on whether or not the pattern matches.
Removed the InvalidVexType error and specific querying of the
vextype value.  This is now handled in the main data loop like all
the rest.  The specific test to ensure it exists is still used.

Fixed some documentation.

Added a function to convert any unicode strings in a data structure
to normal strings.  This replaces many uses of str().