Removed functions to find primitives by name (should just use hou.Geometry.glob[Points/Prims]() to find things matching strings.)
Add mapping support for group attribute owners instead of hardcoded numbers and ternary operators.

Add functions to get the number of instances per multiparm item and the starting offset of a multiparm.

More use of group type hdk mappings.

Add function to check if a parm tuple is a vector and get the values as a corresponding sized hou.Vector object.

Add function to check if a parameter is a color parameter and evaluate the parameter as a color.

Add function to get the referenced hou.Node for operator path parameters.

When querying multiparm instances, properly handle parms and parm tuples based on the tuple size.

Updated tests to reflect removed functions and changed behavior.