Intrinsics are now stored as a list so as to support multiples.  Create/Edit them by using space/command separated words.
When exporting per-category nolonger throw lights with no category set into an imaginary __node__ category, instead just using the unmodified channel name.

Fix a bug where if you attempted to create a new AOV using the manager dialog and save it into an existing file that contained group definitions the groups would lose their list of included aovs.

The New/Edit AOV dialog now has an additional option to specity which components get exported.  Previously if the field was left blank the ROP values were used but this was not clear.  There is now a menu to set the components a) from the rop, b) from the aov in which case the explicit list is used.